requirements.txt
python-telegram-bot==20.8
websockets==11.0
aiohttp==3.8.4